requirements.txt
django-tinymce
six
pillow

python -m pip install django-twilio